+++ basic-sample/src/main/java/cz/trask/zenid/sample/ui/HologramActivity.java
@@ -4,15 +4,19 @@ import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import java.util.Arrays;
+
 import cz.trask.zenid.sample.LogUtils;
 import cz.trask.zenid.sample.R;
+import cz.trask.zenid.sdk.DocumentAcceptableInput;
 import cz.trask.zenid.sdk.DocumentCountry;
 import cz.trask.zenid.sdk.DocumentPage;
-import cz.trask.zenid.sdk.DocumentResult;
 import cz.trask.zenid.sdk.DocumentRole;
+import cz.trask.zenid.sdk.HologramResult;
 import cz.trask.zenid.sdk.HologramState;
 import cz.trask.zenid.sdk.HologramView;
 import cz.trask.zenid.sdk.Language;
+import cz.trask.zenid.sdk.VisualizationSettings;
 import timber.log.Timber;
 
 public class HologramActivity extends AppCompatActivity {
@@ -24,10 +28,19 @@ public class HologramActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_hologram);
 
+        DocumentAcceptableInput.Filter filter1 = new DocumentAcceptableInput.Filter(DocumentRole.ID,  DocumentPage.FRONT_SIDE, DocumentCountry.CZ);
+        DocumentAcceptableInput documentAcceptableInput = new DocumentAcceptableInput(Arrays.asList(filter1));
+
+
+        VisualizationSettings visualizationSettings = new VisualizationSettings.Builder()
+                .showDebugVisualization(true)
+                .language(Language.ENGLISH)
+                .build();
+
         hologramView = findViewById(R.id.hologramView);
-        hologramView.setDocumentType(DocumentRole.ID, DocumentPage.FRONT_SIDE, DocumentCountry.CZ);
         hologramView.setLifecycleOwner(this);
-        hologramView.enableDefaultVizualization(Language.ENGLISH);
+        hologramView.setDocumentAcceptableInput(documentAcceptableInput);
+        hologramView.enableDefaultVisualization(visualizationSettings);
         hologramView.setCallback(new HologramView.Callback() {
 
             @Override
@@ -36,8 +49,8 @@ public class HologramActivity extends AppCompatActivity {
             }
 
             @Override
-            public void onVideoTaken(DocumentResult result) {
-                LogUtils.logInfo(getApplicationContext(), "onVideoTaken... " + result.getFilePath());
+            public void onVideoTaken(HologramResult result) {
+                LogUtils.logInfo(getApplicationContext(), "onVideoTaken... " + result.getVideoFilePath());
             }
         });
     }